Revert "move io.gscontrol_mmix to utils"
This reverts commit 5889135d3459feb9e2d02c815a046bd33ef57505.
