+++ tedana/workflows/tedana.py
@@ -434,7 +434,7 @@ def tedana_workflow(data, tes, mask=None, mixm=None, ctab=None, manacc=None,
 
     io.writeresults(OCcatd, mask, comptable, mmix, fixed_seed, n_vols,
                     acc, rej, midk, empty, ref_img)
-    utils.gscontrol_mmix(OCcatd, mmix, mask, acc, ref_img)
+    io.gscontrol_mmix(OCcatd, mmix, mask, acc, ref_img)
     if dne:
         io.writeresults_echoes(catd, mmix, mask, acc, rej, midk, ref_img)
